chore: took api from 3 files to 1, premature optmization
