--- frontend/src/features/story/StoryPhase/api/story-phase.constants.ts
@@ -1,2 +1,0 @@
-export const PROJECT_ENDPOINT =
-  "/api/comic-builder/v2/projects" as const;
