--- frontend/src/features/story/StoryPhase/api/story-phase.query-keys.ts
@@ -1,5 +1,0 @@
-export const storyPhaseKeys = {
-  all: ["storyPhase"] as const,
-  project: (projectId: string) =>
-    [...storyPhaseKeys.all, "project", projectId] as const,
-};
